[WIP] Rewrite AccountsActivity to M3
