+++ app/src/main/kotlin/at/bitfire/davdroid/ui/AccountsDrawerHandler.kt
@@ -9,34 +9,32 @@ import android.content.Context
 import android.content.Intent
 import androidx.annotation.StringRes
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.Divider
-import androidx.compose.material.Icon
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.SnackbarHostState
-import androidx.compose.material.SnackbarResult
-import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Feedback
 import androidx.compose.material.icons.filled.Info
 import androidx.compose.material.icons.filled.Settings
 import androidx.compose.material.icons.filled.Storage
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.NavigationDrawerItem
+import androidx.compose.material3.SnackbarHostState
+import androidx.compose.material3.SnackbarResult
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.compositionLocalOf
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.scale
 import androidx.compose.ui.graphics.Color
@@ -47,6 +45,7 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalInspectionMode
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.core.net.toUri
 import at.bitfire.davdroid.BuildConfig
@@ -77,7 +76,7 @@ abstract class AccountsDrawerHandler {
         snackbarHostState: SnackbarHostState,
         onCloseDrawer: () -> Unit
     ) {
-        Column(Modifier
+        Column(modifier = Modifier
             .fillMaxWidth()
             .verticalScroll(rememberScrollState())
         ) {
@@ -181,15 +180,14 @@ abstract class AccountsDrawerHandler {
     fun Header() {
         Column(
             Modifier
-            .background(Color.DarkGray)
-            .fillMaxWidth()
-            .padding(16.dp)
+                .background(Color.DarkGray)
+                .fillMaxWidth()
+                .padding(16.dp)
         ) {
             Spacer(Modifier.height(16.dp))
             Box(
-                Modifier
-                .background(
-                    color = MaterialTheme.colors.primary,
+                Modifier.background(
+                    color = MaterialTheme.colorScheme.primary,
                     shape = RoundedCornerShape(16.dp)
                 )
             ) {
@@ -199,8 +197,7 @@ abstract class AccountsDrawerHandler {
                     tint = Color.White,
                     modifier = Modifier
                         .scale(1.2f)
-                        .height(56.dp)
-                        .width(56.dp)
+                        .size(64.dp)
                 )
             }
             Spacer(Modifier.height(8.dp))
@@ -208,12 +205,12 @@ abstract class AccountsDrawerHandler {
             Text(
                 stringResource(R.string.app_name),
                 color = Color.White,
-                style = MaterialTheme.typography.body1
+                style = MaterialTheme.typography.bodyLarge
             )
             Text(
                 stringResource(R.string.navigation_drawer_subtitle),
                 color = Color.White.copy(alpha = 0.7f),
-                style = MaterialTheme.typography.body2
+                style = MaterialTheme.typography.bodyMedium
             )
         }
 
@@ -222,11 +219,11 @@ abstract class AccountsDrawerHandler {
 
     @Composable
     fun MenuHeading(text: String) {
-        Divider(Modifier.padding(vertical = 8.dp))
+        HorizontalDivider(Modifier.padding(vertical = 8.dp))
 
         Text(
             text,
-            style = MaterialTheme.typography.body2,
+            style = MaterialTheme.typography.bodyLarge.copy(fontWeight = FontWeight.Bold),
             modifier = Modifier.padding(8.dp)
         )
     }
@@ -241,28 +238,15 @@ abstract class AccountsDrawerHandler {
         onClick: () -> Unit
     ) {
         val closeHandler = localCloseDrawerHandler.current
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            modifier = Modifier
-                .clickable(onClick = {
-                    onClick()
-                    closeHandler.closeDrawer()
-                })
-                .padding(4.dp)
-        ) {
-            Icon(
-                icon,
-                contentDescription = title,
-                modifier = Modifier.padding(horizontal = 8.dp, vertical = 4.dp)
-            )
-            Text(
-                title,
-                style = MaterialTheme.typography.body2,
-                modifier = Modifier
-                    .padding(start = 8.dp)
-                    .weight(1f)
-            )
-        }
+        NavigationDrawerItem(
+            icon = { Icon(icon, contentDescription = title) },
+            label = { Text(title) },
+            selected = false,
+            onClick = {
+                onClick()
+                closeHandler.closeDrawer()
+            }
+        )
     }
 
     @Composable
