+++ app/src/main/kotlin/at/bitfire/davdroid/ui/OseAccountsDrawerHandler.kt
@@ -5,7 +5,6 @@
 package at.bitfire.davdroid.ui
 
 import androidx.compose.foundation.layout.Column
-import androidx.compose.material.SnackbarHostState
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.HelpCenter
 import androidx.compose.material.icons.filled.CloudOff
@@ -13,6 +12,7 @@ import androidx.compose.material.icons.filled.Forum
 import androidx.compose.material.icons.filled.Home
 import androidx.compose.material.icons.filled.Info
 import androidx.compose.material.icons.filled.VolunteerActivism
+import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.platform.LocalUriHandler
 import androidx.compose.ui.res.painterResource
